Update tests to reflect new module organisation
